+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Changelog
 
+## v1.3.1
+
+### Added
+
 ## v1.3.0
 
 ### Added